refactor(plot): relocate and centralize gen_isi_hist function
- Removed `gen_isi_hist` from `utils/blech_process_utils.py` and added it to `blech_units_plot.py` for better organization and maintainability.
- Updated imports in `blech_post_process_utils.py` and `blech_process_utils.py` to reflect the new location of `gen_isi_hist`.
- Replaced the usage of `gen_isi_hist` within each relevant module with the newly relocated function.
